refactor(storybook): move .storybook config dir into frontend/
The Storybook config is frontend-only, so co-locate it under frontend/
instead of cluttering the repo root.

- git mv .storybook -> frontend/.storybook
- point the dev/build scripts at the new dir via -c frontend/.storybook
- fix config-relative paths in main.ts (stories glob, @lexicons alias)
  and the ../frontend/src imports in fixtures/mockStore/preview
- update the ../../../../.storybook fixture imports in story files

Moving the config dir changes how Storybook's Vite builder resolves the
project root (configDir/..), so it now auto-discovers frontend/vite.config.ts
and its vite-plugin-pwa fails the build on the >2 MiB manager bundle. Pin an
isolated empty Storybook Vite config via builder.viteConfigPath to preserve
the previous isolated-build behavior (the root had no vite config to load).
